+++ packages/engine/src/schemas/projects/project-build.schema.ts
@@ -33,7 +33,8 @@ export const ProjectBuildUpdateItemSchema = Type.Object(
     sourceURL: Type.String(),
     destinationURL: Type.String(),
     name: Type.String(),
-    reset: Type.Boolean(),
+    needsRebuild: Type.Optional(Type.Boolean()),
+    reset: Type.Optional(Type.Boolean()),
     commitSHA: Type.String(),
     sourceBranch: Type.String(),
     updateType: StringEnum(projectUpdateTypes),
